Restructure hl reconstruction module
Reformatted to current style in some places.
Reorganized source code for readability.
