+++ src/iop/highlights.c
@@ -44,10 +44,9 @@
 #include <gtk/gtk.h>
 #include <inttypes.h>
 
-#define MAX_NUM_SCALES 12
-
 // Downsampling factor for guided-laplacian
 #define DS_FACTOR 4
+#define MAX_NUM_SCALES 12
 
 DT_MODULE_INTROSPECTION(4, dt_iop_highlights_params_t)
 
@@ -197,8 +196,11 @@ int default_colorspace(dt_iop_module_t *self, dt_dev_pixelpipe_t *pipe, dt_dev_p
   return IOP_CS_RAW;
 }
 
-int legacy_params(dt_iop_module_t *self, const void *const old_params, const int old_version,
-                  void *new_params, const int new_version)
+int legacy_params(dt_iop_module_t *self,
+                  const void *const old_params,
+                  const int old_version,
+                  void *new_params,
+                  const int new_version)
 {
   if(old_version == 1 && new_version == 4)
   {
@@ -264,17 +266,192 @@ int legacy_params(dt_iop_module_t *self, const void *const old_params, const int
 static dt_aligned_pixel_t img_oppchroma;
 static uint64_t img_opphash = 0;
 
+#include "common/distance_transform.h"
+#include "hlreconstruct/segmentation.c"
+#include "hlreconstruct/segbased.c"
+#include "hlreconstruct/opposed.c"
+#include "hlreconstruct/laplacian.c"
+#include "hlreconstruct/inpaint.c"
+#include "hlreconstruct/lch.c"
+
+/* inpaint opposed and segmentation based algorithms want the whole image for proper calculation
+   of chrominance correction and best candidates so we change both rois.
+*/
+void modify_roi_out(dt_iop_module_t *self,
+                    dt_dev_pixelpipe_iop_t *piece,
+                    dt_iop_roi_t *roi_out,
+                    const dt_iop_roi_t *const roi_in)
+{
+  *roi_out = *roi_in;
+  // it can never hurt to make sure
+  roi_out->x = MAX(0, roi_in->x);
+  roi_out->y = MAX(0, roi_in->y);
+}
+
+void modify_roi_in(dt_iop_module_t *self,
+                   dt_dev_pixelpipe_iop_t *piece,
+                   const dt_iop_roi_t *const roi_out,
+                   dt_iop_roi_t *roi_in)
+{
+  *roi_in = *roi_out;
+
+  dt_iop_highlights_data_t *d = (dt_iop_highlights_data_t *)piece->data;
+  const gboolean use_opposing = (d->mode == DT_IOP_HIGHLIGHTS_OPPOSED) || (d->mode == DT_IOP_HIGHLIGHTS_SEGMENTS);
+  const gboolean fullpipe = piece->pipe->type & DT_DEV_PIXELPIPE_FULL;
+  /* When do we need to expand the roi to maximum of the full input data?
+     1. Certainly not if any other than opposed or the segmentation based algo is used.
+   */
+  if(!use_opposing)
+    return;
+
+  dt_iop_highlights_gui_data_t *g = (dt_iop_highlights_gui_data_t *)self->gui_data;
+  const gboolean clipmask = (g != NULL) ? (g->hlr_mask_mode == DT_HIGHLIGHTS_MASK_CLIPPED) : FALSE;
+  /*
+     2. Certainly not if we show the clipped mask in fullpipe as that is also safe with current roi
+     3. Certainly not for linear raws as they miss the automatic downscaler provided by the demosaicer stage, so
+        the expanding to full image data does not work as we do a downscaling very early in
+        the pixelpipe. So - no quality achieved but really bad performance.
+        See #12998 and #12993 for a lengthy discussion
+  */
+  if((fullpipe && clipmask) || (piece->pipe->dsc.filters == 0))
+    return;
+
+  /* We require the correct (full-image-data) expansion with a defined scale for all pixelpipes for proper
+     aligning and scaling in the demosiacer
+  */
+  roi_in->x = 0;
+  roi_in->y = 0;
+  roi_in->width = piece->buf_in.width;
+  roi_in->height = piece->buf_in.height;
+  roi_in->scale = 1.0f;
+}
+
+void tiling_callback(struct dt_iop_module_t *self,
+                     struct dt_dev_pixelpipe_iop_t *piece,
+                     const dt_iop_roi_t *roi_in,
+                     const dt_iop_roi_t *roi_out,
+                     struct dt_develop_tiling_t *tiling)
+{
+  dt_iop_highlights_data_t *d = (dt_iop_highlights_data_t *)piece->data;
+  const uint32_t filters = piece->pipe->dsc.filters;
+
+  if(d->mode == DT_IOP_HIGHLIGHTS_LAPLACIAN && filters && filters != 9u)
+  {
+    // Bayer CFA and guided laplacian method : prepare for wavelets decomposition.
+    const float scale = fmaxf(DS_FACTOR * piece->iscale / roi_in->scale, 1.f);
+    const float final_radius = (float)((int)(1 << d->scales)) / scale;
+    const int scales = CLAMP((int)ceilf(log2f(final_radius)), 1, MAX_NUM_SCALES);
+    const int max_filter_radius = (1 << scales);
+
+    // Warning : in and out are single-channel in RAW mode
+    // in + out + interpolated + ds_interpolated + ds_tmp + 2 * ds_LF + ds_HF + mask + ds_mask
+    if(filters) // RAW
+    {
+      tiling->factor = 2.f + 2.f * 4 + 6.f * 4 / (DS_FACTOR * DS_FACTOR);
+      tiling->factor_cl =  2.f + 3.f * 4 + 5.f * 4 / (DS_FACTOR * DS_FACTOR);
+
+      // The wavelets decomposition uses a temp buffer of size 4 × ds_width
+      tiling->maxbuf = 1.f / roi_in->height * dt_get_num_threads() * 4.f / DS_FACTOR;
+    }
+    else
+    {
+      tiling->factor = 2.f + 2.f + 6.f / (DS_FACTOR * DS_FACTOR);
+      tiling->factor_cl = 2.f + 3.f + 5.f / (DS_FACTOR * DS_FACTOR);
+
+      // The wavelets decomposition uses a temp buffer of size 4 × width
+      tiling->maxbuf = 1.f / roi_in->height * dt_get_num_threads() / DS_FACTOR;
+    }
+
+    // No temp buffer on GPU
+    tiling->maxbuf_cl = 1.0f;
+    tiling->overhead = 0;
+
+    // Note : if we were not doing anything iterative,
+    // max_filter_radius would not need to be factored more.
+    // Since we are iterating within tiles, we need more padding.
+    // The clean way of doing it would be an internal tiling mechanism
+    // where we restitch the tiles between each new iteration.
+    tiling->overlap = max_filter_radius * 1.5f / DS_FACTOR;
+    tiling->xalign = 1;
+    tiling->yalign = 1;
+
+    return;
+  }
+
+  if(d->mode == DT_IOP_HIGHLIGHTS_SEGMENTS)
+  {
+    // even if the algorithm can't tile we want to calculate memory for pixelpipe checks and a possible warning
+    const int segments = roi_out->width * roi_out->height / 4000; // segments per mpix
+    if(filters == 9u)
+    {
+      tiling->xalign = 3;
+      tiling->yalign = 3;
+    }
+    else
+    {
+      tiling->xalign = 2;
+      tiling->yalign = 2;
+    }
+    tiling->overlap = 0;
+    tiling->overhead = segments * 5 * 5 * sizeof(int); // segmentation stuff
+    tiling->factor = 3.0f;
+    tiling->maxbuf = 1.0f;
+    return;
+  }
+
+  if(d->mode == DT_IOP_HIGHLIGHTS_OPPOSED)
+  {
+    if(filters == 9u)
+    {
+      tiling->xalign = 3;
+      tiling->yalign = 3;
+    }
+    else
+    {
+      tiling->xalign = 2;
+      tiling->yalign = 2;
+    }
+    tiling->factor = 2.5f; // enough for in&output buffers plus masks
+    tiling->overlap = 0;
+    tiling->maxbuf = 1.0f;
+    tiling->overhead = 0;
+    return;
+  }
+
+  tiling->factor = 2.0f;  // in + out
+  tiling->maxbuf = 1.0f;
+  tiling->overhead = 0;
+
+  if(filters == 9u)
+  {
+    // xtrans
+    tiling->xalign = 6;
+    tiling->yalign = 6;
+    tiling->overlap = (d->mode == DT_IOP_HIGHLIGHTS_LCH) ? 2 : 0;
+  }
+  else if(filters)
+  {
+    // bayer
+    tiling->xalign = 2;
+    tiling->yalign = 2;
+    tiling->overlap = (d->mode == DT_IOP_HIGHLIGHTS_LCH) ? 1 : 0;
+  }
+  else
+  {
+    // non-raw
+    tiling->xalign = 1;
+    tiling->yalign = 1;
+    tiling->overlap = 0;
+  }
+}
+
 #ifdef HAVE_OPENCL
-static cl_int process_laplacian_bayer_cl(struct dt_iop_module_t *self, dt_dev_pixelpipe_iop_t *piece,
-                                         cl_mem dev_in, cl_mem dev_out, const dt_iop_roi_t *const roi_in,
-                                         const dt_iop_roi_t *const roi_out, const dt_aligned_pixel_t clips);
-
-static cl_int process_opposed_cl(struct dt_iop_module_t *self, dt_dev_pixelpipe_iop_t *piece,
-                                         cl_mem dev_in, cl_mem dev_out, const dt_iop_roi_t *const roi_in,
-                                         const dt_iop_roi_t *const roi_out);
-
-int process_cl(struct dt_iop_module_t *self, dt_dev_pixelpipe_iop_t *piece, cl_mem dev_in, cl_mem dev_out,
-               const dt_iop_roi_t *const roi_in, const dt_iop_roi_t *const roi_out)
+int process_cl(struct dt_iop_module_t *self,
+               dt_dev_pixelpipe_iop_t *piece,
+               cl_mem dev_in,
+               cl_mem dev_out,
+               const dt_iop_roi_t *const roi_in,
+               const dt_iop_roi_t *const roi_out)
 {
   dt_iop_highlights_data_t *d = (dt_iop_highlights_data_t *)piece->data;
   dt_iop_highlights_gui_data_t *g = (dt_iop_highlights_gui_data_t *)self->gui_data;
@@ -425,1533 +602,11 @@ int process_cl(struct dt_iop_module_t *self, dt_dev_pixelpipe_iop_t *piece, cl_m
 }
 #endif
 
-void tiling_callback(struct dt_iop_module_t *self, struct dt_dev_pixelpipe_iop_t *piece,
-              const dt_iop_roi_t *roi_in, const dt_iop_roi_t *roi_out,
-              struct dt_develop_tiling_t *tiling)
-{
-  dt_iop_highlights_data_t *d = (dt_iop_highlights_data_t *)piece->data;
-  const uint32_t filters = piece->pipe->dsc.filters;
-
-  if(d->mode == DT_IOP_HIGHLIGHTS_LAPLACIAN && filters && filters != 9u)
-  {
-    // Bayer CFA and guided laplacian method : prepare for wavelets decomposition.
-    const float scale = fmaxf(DS_FACTOR * piece->iscale / roi_in->scale, 1.f);
-    const float final_radius = (float)((int)(1 << d->scales)) / scale;
-    const int scales = CLAMP((int)ceilf(log2f(final_radius)), 1, MAX_NUM_SCALES);
-    const int max_filter_radius = (1 << scales);
-
-    // Warning : in and out are single-channel in RAW mode
-    // in + out + interpolated + ds_interpolated + ds_tmp + 2 * ds_LF + ds_HF + mask + ds_mask
-    if(filters) // RAW
-    {
-      tiling->factor = 2.f + 2.f * 4 + 6.f * 4 / (DS_FACTOR * DS_FACTOR);
-      tiling->factor_cl =  2.f + 3.f * 4 + 5.f * 4 / (DS_FACTOR * DS_FACTOR);
-
-      // The wavelets decomposition uses a temp buffer of size 4 × ds_width
-      tiling->maxbuf = 1.f / roi_in->height * dt_get_num_threads() * 4.f / DS_FACTOR;
-    }
-    else
-    {
-      tiling->factor = 2.f + 2.f + 6.f / (DS_FACTOR * DS_FACTOR);
-      tiling->factor_cl = 2.f + 3.f + 5.f / (DS_FACTOR * DS_FACTOR);
-
-      // The wavelets decomposition uses a temp buffer of size 4 × width
-      tiling->maxbuf = 1.f / roi_in->height * dt_get_num_threads() / DS_FACTOR;
-    }
-
-    // No temp buffer on GPU
-    tiling->maxbuf_cl = 1.0f;
-    tiling->overhead = 0;
-
-    // Note : if we were not doing anything iterative,
-    // max_filter_radius would not need to be factored more.
-    // Since we are iterating within tiles, we need more padding.
-    // The clean way of doing it would be an internal tiling mechanism
-    // where we restitch the tiles between each new iteration.
-    tiling->overlap = max_filter_radius * 1.5f / DS_FACTOR;
-    tiling->xalign = 1;
-    tiling->yalign = 1;
-
-    return;
-  }
-
-  if(d->mode == DT_IOP_HIGHLIGHTS_SEGMENTS)
-  {
-    // even if the algorithm can't tile we want to calculate memory for pixelpipe checks and a possible warning
-    const int segments = roi_out->width * roi_out->height / 4000; // segments per mpix
-    if(filters == 9u)
-    {
-      tiling->xalign = 3;
-      tiling->yalign = 3;
-    }
-    else
-    {
-      tiling->xalign = 2;
-      tiling->yalign = 2;
-    }
-    tiling->overlap = 0;
-    tiling->overhead = segments * 5 * 5 * sizeof(int); // segmentation stuff
-    tiling->factor = 3.0f;
-    tiling->maxbuf = 1.0f;
-    return;
-  }
-
-  if(d->mode == DT_IOP_HIGHLIGHTS_OPPOSED)
-  {
-    if(filters == 9u)
-    {
-      tiling->xalign = 3;
-      tiling->yalign = 3;
-    }
-    else
-    {
-      tiling->xalign = 2;
-      tiling->yalign = 2;
-    }
-    tiling->factor = 2.5f; // enough for in&output buffers plus masks
-    tiling->overlap = 0;
-    tiling->maxbuf = 1.0f;
-    tiling->overhead = 0;
-    return;
-  }
-
-  tiling->factor = 2.0f;  // in + out
-  tiling->maxbuf = 1.0f;
-  tiling->overhead = 0;
-
-  if(filters == 9u)
-  {
-    // xtrans
-    tiling->xalign = 6;
-    tiling->yalign = 6;
-    tiling->overlap = (d->mode == DT_IOP_HIGHLIGHTS_LCH) ? 2 : 0;
-  }
-  else if(filters)
-  {
-    // bayer
-    tiling->xalign = 2;
-    tiling->yalign = 2;
-    tiling->overlap = (d->mode == DT_IOP_HIGHLIGHTS_LCH) ? 1 : 0;
-  }
-  else
-  {
-    // non-raw
-    tiling->xalign = 1;
-    tiling->yalign = 1;
-    tiling->overlap = 0;
-  }
-}
-
-/* interpolate value for a pixel, ideal via ratio to nearby pixel */
-static inline float interp_pix_xtrans(const int ratio_next,
-                                      const ssize_t offset_next,
-                                      const float clip0, const float clip_next,
-                                      const float *const in,
-                                      const float *const ratios)
-{
-  assert(ratio_next != 0);
-  // it's OK to exceed clipping of current pixel's color based on a
-  // neighbor -- that is the purpose of interpolating highlight
-  // colors
-  const float clip_val = fmaxf(clip0, clip_next);
-  if(in[offset_next] >= clip_next - 1e-5f)
-  {
-    // next pixel is also clipped
-    return clip_val;
-  }
-  else
-  {
-    // set this pixel in ratio to the next
-    assert(ratio_next != 0);
-    if(ratio_next > 0)
-      return fminf(in[offset_next] / ratios[ratio_next], clip_val);
-    else
-      return fminf(in[offset_next] * ratios[-ratio_next], clip_val);
-  }
-}
-
-static inline void interpolate_color_xtrans(const void *const ivoid, void *const ovoid,
-                                            const dt_iop_roi_t *const roi_in,
-                                            const dt_iop_roi_t *const roi_out,
-                                            int dim, int dir, int other,
-                                            const float *const clip,
-                                            const uint8_t (*const xtrans)[6],
-                                            const int pass)
-{
-  // In Bayer each row/col has only green/red or green/blue
-  // transitions, hence can reconstruct color by single ratio per
-  // row. In x-trans there can be transitions between arbitrary colors
-  // in a row/col (and 2x2 green blocks which provide no color
-  // transition information). Hence calculate multiple color ratios
-  // for each row/col.
-
-  // Lookup for color ratios, e.g. red -> blue is roff[0][2] and blue
-  // -> red is roff[2][0]. Returned value is an index into ratios. If
-  // negative, then need to invert the ratio. Identity color
-  // transitions aren't used.
-  const int roff[3][3] = {{ 0, -1, -2},
-                          { 1,  0, -3},
-                          { 2,  3,  0}};
-  // record ratios of color transitions 0:unused, 1:RG, 2:RB, and 3:GB
-  dt_aligned_pixel_t ratios = {1.0f, 1.0f, 1.0f, 1.0f};
-
-  // passes are 0:+x, 1:-x, 2:+y, 3:-y
-  // dims are 0:traverse a row, 1:traverse a column
-  // dir is 1:left to right, -1: right to left
-  int i = (dim == 0) ? 0 : other;
-  int j = (dim == 0) ? other : 0;
-  const ssize_t offs = (ssize_t)(dim ? roi_out->width : 1) * ((dir < 0) ? -1 : 1);
-  const ssize_t offl = offs - (dim ? 1 : roi_out->width);
-  const ssize_t offr = offs + (dim ? 1 : roi_out->width);
-  int beg, end;
-  if(dir == 1)
-  {
-    beg = 0;
-    end = (dim == 0) ? roi_out->width : roi_out->height;
-  }
-  else
-  {
-    beg = ((dim == 0) ? roi_out->width : roi_out->height) - 1;
-    end = -1;
-  }
-
-  float *in, *out;
-  if(dim == 1)
-  {
-    out = (float *)ovoid + (size_t)i + (size_t)beg * roi_out->width;
-    in = (float *)ivoid + (size_t)i + (size_t)beg * roi_in->width;
-  }
-  else
-  {
-    out = (float *)ovoid + (size_t)beg + (size_t)j * roi_out->width;
-    in = (float *)ivoid + (size_t)beg + (size_t)j * roi_in->width;
-  }
-
-  for(int k = beg; k != end; k += dir)
-  {
-    if(dim == 1)
-      j = k;
-    else
-      i = k;
-
-    const uint8_t f0 = FCxtrans(j, i, roi_in, xtrans);
-    const uint8_t f1 = FCxtrans(dim ? (j + dir) : j, dim ? i : (i + dir), roi_in, xtrans);
-    const uint8_t fl = FCxtrans(dim ? (j + dir) : (j - 1), dim ? (i - 1) : (i + dir), roi_in, xtrans);
-    const uint8_t fr = FCxtrans(dim ? (j + dir) : (j + 1), dim ? (i + 1) : (i + dir), roi_in, xtrans);
-    const float clip0 = clip[f0];
-    const float clip1 = clip[f1];
-    const float clipl = clip[fl];
-    const float clipr = clip[fr];
-    const float clip_max = fmaxf(fmaxf(clip[0], clip[1]), clip[2]);
-
-    if(i == 0 || i == roi_out->width - 1 || j == 0 || j == roi_out->height - 1)
-    {
-      if(pass == 3) out[0] = fminf(clip_max, in[0]);
-    }
-    else
-    {
-      // ratio to next pixel if this & next are unclamped and not in
-      // 2x2 green block
-      if((f0 != f1) &&
-          (in[0] < clip0 && in[0] > 1e-5f) &&
-          (in[offs] < clip1 && in[offs] > 1e-5f))
-      {
-        const int r = roff[f0][f1];
-        assert(r != 0);
-        if(r > 0)
-          ratios[r] = (3.f * ratios[r] + (in[offs] / in[0])) / 4.f;
-        else
-          ratios[-r] = (3.f * ratios[-r] + (in[0] / in[offs])) / 4.f;
-      }
-
-      if(in[0] >= clip0 - 1e-5f)
-      {
-        // interplate color for clipped pixel
-        float add;
-        if(f0 != f1)
-          // next pixel is different color
-          add =
-            interp_pix_xtrans(roff[f0][f1], offs, clip0, clip1, in, ratios);
-        else
-          // at start of 2x2 green block, look diagonally
-          add = (fl != f0) ?
-            interp_pix_xtrans(roff[f0][fl], offl, clip0, clipl, in, ratios) :
-            interp_pix_xtrans(roff[f0][fr], offr, clip0, clipr, in, ratios);
-
-        if(pass == 0)
-          out[0] = add;
-        else if(pass == 3)
-          out[0] = fminf(clip_max, (out[0] + add) / 4.0f);
-        else
-          out[0] += add;
-      }
-      else
-      {
-        // pixel is not clipped
-        if(pass == 3) out[0] = in[0];
-      }
-    }
-    out += offs;
-    in += offs;
-  }
-}
-
-static inline void interpolate_color(const void *const ivoid, void *const ovoid,
-                                     const dt_iop_roi_t *const roi_out, int dim, int dir, int other,
-                                     const float *clip, const uint32_t filters, const int pass)
-{
-  float ratio = 1.0f;
-  float *in, *out;
-
-  int i = 0, j = 0;
-  if(dim == 0)
-    j = other;
-  else
-    i = other;
-  ssize_t offs = dim ? roi_out->width : 1;
-  if(dir < 0) offs = -offs;
-  int beg, end;
-  if(dim == 0 && dir == 1)
-  {
-    beg = 0;
-    end = roi_out->width;
-  }
-  else if(dim == 0 && dir == -1)
-  {
-    beg = roi_out->width - 1;
-    end = -1;
-  }
-  else if(dim == 1 && dir == 1)
-  {
-    beg = 0;
-    end = roi_out->height;
-  }
-  else if(dim == 1 && dir == -1)
-  {
-    beg = roi_out->height - 1;
-    end = -1;
-  }
-  else
-    return;
-
-  if(dim == 1)
-  {
-    out = (float *)ovoid + i + (size_t)beg * roi_out->width;
-    in = (float *)ivoid + i + (size_t)beg * roi_out->width;
-  }
-  else
-  {
-    out = (float *)ovoid + beg + (size_t)j * roi_out->width;
-    in = (float *)ivoid + beg + (size_t)j * roi_out->width;
-  }
-  for(int k = beg; k != end; k += dir)
-  {
-    if(dim == 1)
-      j = k;
-    else
-      i = k;
-    const float clip0 = clip[FC(j, i, filters)];
-    const float clip1 = clip[FC(dim ? (j + 1) : j, dim ? i : (i + 1), filters)];
-    if(i == 0 || i == roi_out->width - 1 || j == 0 || j == roi_out->height - 1)
-    {
-      if(pass == 3) out[0] = in[0];
-    }
-    else
-    {
-      if(in[0] < clip0 && in[0] > 1e-5f)
-      { // both are not clipped
-        if(in[offs] < clip1 && in[offs] > 1e-5f)
-        { // update ratio, exponential decay. ratio = in[odd]/in[even]
-          if(k & 1)
-            ratio = (3.0f * ratio + in[0] / in[offs]) / 4.0f;
-          else
-            ratio = (3.0f * ratio + in[offs] / in[0]) / 4.0f;
-        }
-      }
-
-      if(in[0] >= clip0 - 1e-5f)
-      { // in[0] is clipped, restore it as in[1] adjusted according to ratio
-        float add = 0.0f;
-        if(in[offs] >= clip1 - 1e-5f)
-          add = fmaxf(clip0, clip1);
-        else if(k & 1)
-          add = in[offs] * ratio;
-        else
-          add = in[offs] / ratio;
-
-        if(pass == 0)
-          out[0] = add;
-        else if(pass == 3)
-          out[0] = (out[0] + add) / 4.0f;
-        else
-          out[0] += add;
-      }
-      else
-      {
-        if(pass == 3) out[0] = in[0];
-      }
-    }
-    out += offs;
-    in += offs;
-  }
-}
-
-/*
- * these 2 constants were computed using following Sage code:
- *
- * sqrt3 = sqrt(3)
- * sqrt12 = sqrt(12) # 2*sqrt(3)
- *
- * print 'sqrt3 = ', sqrt3, ' ~= ', RealField(128)(sqrt3)
- * print 'sqrt12 = ', sqrt12, ' ~= ', RealField(128)(sqrt12)
- */
-#define SQRT3 1.7320508075688772935274463415058723669L
-#define SQRT12 3.4641016151377545870548926830117447339L // 2*SQRT3
-
-static void process_lch_bayer(dt_iop_module_t *self, dt_dev_pixelpipe_iop_t *piece, const void *const ivoid,
-                              void *const ovoid, const dt_iop_roi_t *const roi_in,
-                              const dt_iop_roi_t *const roi_out, const float clip)
-{
-  const uint32_t filters = piece->pipe->dsc.filters;
-
-#ifdef _OPENMP
-#pragma omp parallel for default(none) \
-  dt_omp_firstprivate(clip, filters, ivoid, ovoid, roi_out) \
-  schedule(static) collapse(2)
-#endif
-  for(int j = 0; j < roi_out->height; j++)
-  {
-    for(int i = 0; i < roi_out->width; i++)
-    {
-      float *const out = (float *)ovoid + (size_t)roi_out->width * j + i;
-      const float *const in = (float *)ivoid + (size_t)roi_out->width * j + i;
-
-      if(i == roi_out->width - 1 || j == roi_out->height - 1)
-      {
-        // fast path for border
-        out[0] = MIN(clip, in[0]);
-      }
-      else
-      {
-        int clipped = 0;
-
-        // sample 1 bayer block. thus we will have 2 green values.
-        float R = 0.0f, Gmin = FLT_MAX, Gmax = -FLT_MAX, B = 0.0f;
-        for(int jj = 0; jj <= 1; jj++)
-        {
-          for(int ii = 0; ii <= 1; ii++)
-          {
-            const float val = in[(size_t)jj * roi_out->width + ii];
-
-            clipped = (clipped || (val > clip));
-
-            const int c = FC(j + jj + roi_out->y, i + ii + roi_out->x, filters);
-            switch(c)
-            {
-              case 0:
-                R = val;
-                break;
-              case 1:
-                Gmin = MIN(Gmin, val);
-                Gmax = MAX(Gmax, val);
-                break;
-              case 2:
-                B = val;
-                break;
-            }
-          }
-        }
-
-        if(clipped)
-        {
-          const float Ro = MIN(R, clip);
-          const float Go = MIN(Gmin, clip);
-          const float Bo = MIN(B, clip);
-
-          const float L = (R + Gmax + B) / 3.0f;
-
-          float C = SQRT3 * (R - Gmax);
-          float H = 2.0f * B - Gmax - R;
-
-          const float Co = SQRT3 * (Ro - Go);
-          const float Ho = 2.0f * Bo - Go - Ro;
-
-          if(R != Gmax && Gmax != B)
-          {
-            const float ratio = sqrtf((Co * Co + Ho * Ho) / (C * C + H * H));
-            C *= ratio;
-            H *= ratio;
-          }
-
-          dt_aligned_pixel_t RGB = { 0.0f, 0.0f, 0.0f };
-
-          /*
-           * backtransform proof, sage:
-           *
-           * R,G,B,L,C,H = var('R,G,B,L,C,H')
-           * solve([L==(R+G+B)/3, C==sqrt(3)*(R-G), H==2*B-G-R], R, G, B)
-           *
-           * result:
-           * [[R == 1/6*sqrt(3)*C - 1/6*H + L, G == -1/6*sqrt(3)*C - 1/6*H + L, B == 1/3*H + L]]
-           */
-          RGB[0] = L - H / 6.0f + C / SQRT12;
-          RGB[1] = L - H / 6.0f - C / SQRT12;
-          RGB[2] = L + H / 3.0f;
-
-          out[0] = RGB[FC(j + roi_out->y, i + roi_out->x, filters)];
-        }
-        else
-        {
-          out[0] = in[0];
-        }
-      }
-    }
-  }
-}
-
-static void process_lch_xtrans(dt_iop_module_t *self, dt_dev_pixelpipe_iop_t *piece, const void *const ivoid,
-                               void *const ovoid, const dt_iop_roi_t *const roi_in,
-                               const dt_iop_roi_t *const roi_out, const float clip)
-{
-  const uint8_t(*const xtrans)[6] = (const uint8_t(*const)[6])piece->pipe->dsc.xtrans;
-
-#ifdef _OPENMP
-#pragma omp parallel for default(none) \
-  dt_omp_firstprivate(clip, ivoid, ovoid, roi_in, roi_out, xtrans) \
-  schedule(static)
-#endif
-  for(int j = 0; j < roi_out->height; j++)
-  {
-    float *out = (float *)ovoid + (size_t)roi_out->width * j;
-    float *in = (float *)ivoid + (size_t)roi_in->width * j;
-
-    // bit vector used as ring buffer to remember clipping of current
-    // and last two columns, checking current pixel and its vertical
-    // neighbors
-    int cl = 0;
-
-    for(int i = 0; i < roi_out->width; i++)
-    {
-      // update clipping ring buffer
-      cl = (cl << 1) & 6;
-      if(j >= 2 && j <= roi_out->height - 3)
-      {
-        cl |= (in[-roi_in->width] > clip) | (in[0] > clip) | (in[roi_in->width] > clip);
-      }
-
-      if(i < 2 || i > roi_out->width - 3 || j < 2 || j > roi_out->height - 3)
-      {
-        // fast path for border
-        out[0] = MIN(clip, in[0]);
-      }
-      else
-      {
-        // if current pixel is clipped, always reconstruct
-        int clipped = (in[0] > clip);
-        if(!clipped)
-        {
-          clipped = cl;
-          if(clipped)
-          {
-            // If the ring buffer can't show we are in an obviously
-            // unclipped region, this is the slow case: check if there
-            // is any 3x3 block touching the current pixel which has
-            // no clipping, as then don't need to reconstruct the
-            // current pixel. This avoids zippering in edge
-            // transitions from clipped to unclipped areas. The
-            // X-Trans sensor seems prone to this, unlike Bayer, due
-            // to its irregular pattern.
-            for(int offset_j = -2; offset_j <= 0; offset_j++)
-            {
-              for(int offset_i = -2; offset_i <= 0; offset_i++)
-              {
-                if(clipped)
-                {
-                  clipped = 0;
-                  for(int jj = offset_j; jj <= offset_j + 2; jj++)
-                  {
-                    for(int ii = offset_i; ii <= offset_i + 2; ii++)
-                    {
-                      const float val = in[(ssize_t)jj * roi_in->width + ii];
-                      clipped = (clipped || (val > clip));
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-
-        if(clipped)
-        {
-          dt_aligned_pixel_t mean = { 0.0f, 0.0f, 0.0f };
-          dt_aligned_pixel_t RGBmax = { -FLT_MAX, -FLT_MAX, -FLT_MAX };
-          int cnt[3] = { 0, 0, 0 };
-
-          for(int jj = -1; jj <= 1; jj++)
-          {
-            for(int ii = -1; ii <= 1; ii++)
-            {
-              const float val = in[(ssize_t)jj * roi_in->width + ii];
-              const int c = FCxtrans(j+jj, i+ii, roi_in, xtrans);
-              mean[c] += val;
-              cnt[c]++;
-              RGBmax[c] = MAX(RGBmax[c], val);
-            }
-          }
-
-          const float Ro = MIN(mean[0]/cnt[0], clip);
-          const float Go = MIN(mean[1]/cnt[1], clip);
-          const float Bo = MIN(mean[2]/cnt[2], clip);
-
-          const float R = RGBmax[0];
-          const float G = RGBmax[1];
-          const float B = RGBmax[2];
-
-          const float L = (R + G + B) / 3.0f;
-
-          float C = SQRT3 * (R - G);
-          float H = 2.0f * B - G - R;
-
-          const float Co = SQRT3 * (Ro - Go);
-          const float Ho = 2.0f * Bo - Go - Ro;
-
-          if(R != G && G != B)
-          {
-            const float ratio = sqrtf((Co * Co + Ho * Ho) / (C * C + H * H));
-            C *= ratio;
-            H *= ratio;
-          }
-
-          dt_aligned_pixel_t RGB = { 0.0f, 0.0f, 0.0f };
-
-          RGB[0] = L - H / 6.0f + C / SQRT12;
-          RGB[1] = L - H / 6.0f - C / SQRT12;
-          RGB[2] = L + H / 3.0f;
-
-          out[0] = RGB[FCxtrans(j, i, roi_out, xtrans)];
-        }
-        else
-          out[0] = in[0];
-      }
-      out++;
-      in++;
-    }
-  }
-}
-
-#undef SQRT3
-#undef SQRT12
-
-static void _interpolate_and_mask(const float *const restrict input,
-                                  float *const restrict interpolated,
-                                  float *const restrict clipping_mask,
-                                  const dt_aligned_pixel_t clips,
-                                  const dt_aligned_pixel_t wb,
-                                  const uint32_t filters,
-                                  const size_t width, const size_t height)
-{
-  // Bilinear interpolation
-  #ifdef _OPENMP
-  #pragma omp parallel for default(none) \
-    dt_omp_firstprivate(width, height, clips, filters, wb)  \
-    dt_omp_sharedconst(input, interpolated, clipping_mask) \
-    schedule(static)
-  #endif
-  for(size_t i = 0; i < height; i++)
-    for(size_t j = 0; j < width; j++)
-    {
-      const size_t c = FC(i, j, filters);
-      const size_t i_center = i * width;
-      const float center = input[i_center + j];
-
-      float R = 0.f;
-      float G = 0.f;
-      float B = 0.f;
-
-      int R_clipped = 0;
-      int G_clipped = 0;
-      int B_clipped = 0;
-
-      if(i == 0 || j == 0 || i == height - 1 || j == width - 1)
-      {
-        // We are on the image edges. We don't need to demosaic,
-        // just set R = G = B = center and record clipping.
-        // This will introduce a marginal error close to edges, mostly irrelevant
-        // because we are dealing with local averages anyway, later on.
-        // Also we remosaic the image at the end, so only the relevant channel gets picked.
-        // Finally, it's unlikely that the borders of the image get clipped due to vignetting.
-        R = G = B = center;
-        R_clipped = G_clipped = B_clipped = (center > clips[c]);
-      }
-      else
-      {
-        const size_t i_prev = (i - 1) * width;
-        const size_t i_next = (i + 1) * width;
-        const size_t j_prev = (j - 1);
-        const size_t j_next = (j + 1);
-
-        const float north = input[i_prev + j];
-        const float south = input[i_next + j];
-        const float west = input[i_center + j_prev];
-        const float east = input[i_center + j_next];
-
-        const float north_east = input[i_prev + j_next];
-        const float north_west = input[i_prev + j_prev];
-        const float south_east = input[i_next + j_next];
-        const float south_west = input[i_next + j_prev];
-
-        if(c == GREEN) // green pixel
-        {
-          G = center;
-          G_clipped = (center > clips[GREEN]);
-        }
-        else // non-green pixel
-        {
-          // interpolate inside an X/Y cross
-          G = (north + south + east + west) / 4.f;
-          G_clipped = (north > clips[GREEN] || south > clips[GREEN] || east > clips[GREEN] || west > clips[GREEN]);
-        }
-
-        if(c == RED ) // red pixel
-        {
-          R = center;
-          R_clipped = (center > clips[RED]);
-        }
-        else // non-red pixel
-        {
-          if(FC(i - 1, j, filters) == RED && FC(i + 1, j, filters) == RED)
-          {
-            // we are on a red column, so interpolate column-wise
-            R = (north + south) / 2.f;
-            R_clipped = (north > clips[RED] || south > clips[RED]);
-          }
-          else if(FC(i, j - 1, filters) == RED && FC(i, j + 1, filters) == RED)
-          {
-            // we are on a red row, so interpolate row-wise
-            R = (west + east) / 2.f;
-            R_clipped = (west > clips[RED] || east > clips[RED]);
-          }
-          else
-          {
-            // we are on a blue row, so interpolate inside a square
-            R = (north_west + north_east + south_east + south_west) / 4.f;
-            R_clipped = (north_west > clips[RED] || north_east > clips[RED] || south_west > clips[RED]
-                          || south_east > clips[RED]);
-          }
-        }
-
-        if(c == BLUE ) // blue pixel
-        {
-          B = center;
-          B_clipped = (center > clips[BLUE]);
-        }
-        else // non-blue pixel
-        {
-          if(FC(i - 1, j, filters) == BLUE && FC(i + 1, j, filters) == BLUE)
-          {
-            // we are on a blue column, so interpolate column-wise
-            B = (north + south) / 2.f;
-            B_clipped = (north > clips[BLUE] || south > clips[BLUE]);
-          }
-          else if(FC(i, j - 1, filters) == BLUE && FC(i, j + 1, filters) == BLUE)
-          {
-            // we are on a red row, so interpolate row-wise
-            B = (west + east) / 2.f;
-            B_clipped = (west > clips[BLUE] || east > clips[BLUE]);
-          }
-          else
-          {
-            // we are on a red row, so interpolate inside a square
-            B = (north_west + north_east + south_east + south_west) / 4.f;
-
-            B_clipped = (north_west > clips[BLUE] || north_east > clips[BLUE] || south_west > clips[BLUE]
-                        || south_east > clips[BLUE]);
-          }
-        }
-      }
-
-      dt_aligned_pixel_t RGB = { R, G, B, sqrtf(sqf(R) + sqf(G) + sqf(B)) };
-      dt_aligned_pixel_t clipped = { R_clipped, G_clipped, B_clipped, (R_clipped || G_clipped || B_clipped) };
-
-      for_each_channel(k, aligned(RGB, interpolated, clipping_mask, clipped, wb))
-      {
-        const size_t idx = (i * width + j) * 4 + k;
-        interpolated[idx] = fmaxf(RGB[k] / wb[k], 0.f);
-        clipping_mask[idx] = clipped[k];
-      }
-    }
-}
-
-static void _remosaic_and_replace(const float *const restrict input,
-                                  const float *const restrict interpolated,
-                                  const float *const restrict clipping_mask,
-                                  float *const restrict output,
-                                  const dt_aligned_pixel_t wb,
-                                  const uint32_t filters,
-                                  const size_t width,
-                                  const size_t height)
-{
-  // Take RGB ratios and norm, reconstruct RGB and remosaic the image
-  #ifdef _OPENMP
-  #pragma omp parallel for default(none) \
-    dt_omp_firstprivate(width, height, filters, wb)  \
-    dt_omp_sharedconst(output, interpolated, input, clipping_mask) \
-    schedule(static)
-  #endif
-  for(size_t i = 0; i < height; i++)
-    for(size_t j = 0; j < width; j++)
-    {
-      const size_t c = FC(i, j, filters);
-      const size_t idx = i * width + j;
-      const size_t index = idx * 4;
-      const float opacity = clipping_mask[index + ALPHA];
-      output[idx] = opacity * fmaxf(interpolated[index + c] * wb[c], 0.f)
-                    + (1.f - opacity) * input[idx];
-    }
-}
-
-typedef enum diffuse_reconstruct_variant_t
-{
-  DIFFUSE_RECONSTRUCT_RGB = 0,
-  DIFFUSE_RECONSTRUCT_CHROMA
-} diffuse_reconstruct_variant_t;
-
-
-enum wavelets_scale_t
-{
-  ANY_SCALE   = 1 << 0, // any wavelets scale   : reconstruct += HF
-  FIRST_SCALE = 1 << 1, // first wavelets scale : reconstruct = 0
-  LAST_SCALE  = 1 << 2, // last wavelets scale  : reconstruct += residual
-};
-
-
-static unsigned int scale_type(const int s, const int scales)
-{
-  unsigned int scale = ANY_SCALE;
-  if(s == 0) scale |= FIRST_SCALE;
-  if(s == scales - 1) scale |= LAST_SCALE;
-  return scale;
-}
-
-
-static inline void guide_laplacians(const float *const restrict high_freq,
-                                    const float *const restrict low_freq,
-                                    const float *const restrict clipping_mask,
-                                    float *const restrict output,
-                                    const size_t width,
-                                    const size_t height,
-                                    const int mult,
-                                    const float noise_level,
-                                    const int salt,
-                                    const unsigned int scale,
-                                    const float radius_sq)
-{
-  float *const restrict out = DT_IS_ALIGNED(output);
-  const float *const restrict LF = DT_IS_ALIGNED(low_freq);
-  const float *const restrict HF = DT_IS_ALIGNED(high_freq);
-
-#ifdef _OPENMP
-#pragma omp parallel for default(none)                                                                            \
-    dt_omp_firstprivate(out, clipping_mask, HF, LF, height, width, mult, noise_level, salt, scale, radius_sq) \
-    schedule(static)
-#endif
-  for(size_t row = 0; row < height; ++row)
-  {
-    // interleave the order in which we process the rows so that we minimize cache misses
-    const int i = dwt_interleave_rows(row, height, mult);
-    // compute the 'above' and 'below' coordinates, clamping them to the image, once for the entire row
-    const size_t i_neighbours[3]
-      = { MAX((int)(i - mult), (int)0) * width,            // x - mult
-          i * width,                                       // x
-          MIN((int)(i + mult), (int)height - 1) * width }; // x + mult
-    for(int j = 0; j < width; ++j)
-    {
-      const size_t idx = (i * width + j);
-      const size_t index = idx * 4;
-
-      // fetch the clipping mask opacity : opaque (alpha = 100 %) where clipped
-      const float alpha = clipping_mask[index + ALPHA];
-      const float alpha_comp = 1.f - clipping_mask[index + ALPHA];
-
-      dt_aligned_pixel_t high_frequency = { HF[index + 0], HF[index + 1], HF[index + 2], HF[index + 3] };
-
-      if(alpha > 0.f) // reconstruct
-      {
-        // non-local neighbours coordinates
-        const size_t j_neighbours[3]
-          = { MAX((int)(j - mult), (int)0),           // y - mult
-              j,                                      // y
-              MIN((int)(j + mult), (int)width - 1) }; // y + mult
-
-        // fetch non-local pixels and store them locally and contiguously
-        dt_aligned_pixel_t neighbour_pixel_HF[9];
-        for_four_channels(c, aligned(neighbour_pixel_HF, HF))
-        {
-          neighbour_pixel_HF[3 * 0 + 0][c] = HF[4 * (i_neighbours[0] + j_neighbours[0]) + c];
-          neighbour_pixel_HF[3 * 0 + 1][c] = HF[4 * (i_neighbours[0] + j_neighbours[1]) + c];
-          neighbour_pixel_HF[3 * 0 + 2][c] = HF[4 * (i_neighbours[0] + j_neighbours[2]) + c];
-
-          neighbour_pixel_HF[3 * 1 + 0][c] = HF[4 * (i_neighbours[1] + j_neighbours[0]) + c];
-          neighbour_pixel_HF[3 * 1 + 1][c] = HF[4 * (i_neighbours[1] + j_neighbours[1]) + c];
-          neighbour_pixel_HF[3 * 1 + 2][c] = HF[4 * (i_neighbours[1] + j_neighbours[2]) + c];
-
-          neighbour_pixel_HF[3 * 2 + 0][c] = HF[4 * (i_neighbours[2] + j_neighbours[0]) + c];
-          neighbour_pixel_HF[3 * 2 + 1][c] = HF[4 * (i_neighbours[2] + j_neighbours[1]) + c];
-          neighbour_pixel_HF[3 * 2 + 2][c] = HF[4 * (i_neighbours[2] + j_neighbours[2]) + c];
-        }
-
-        // Compute the linear fit of the laplacian of chromaticity against the laplacian of the norm
-        // that is the chromaticity filter guided by the norm
-
-        // Get the local average per channel
-        dt_aligned_pixel_t means_HF = { 0.f, 0.f, 0.f, 0.f };
-        for(size_t k = 0; k < 9; k++)
-          for_each_channel(c, aligned(neighbour_pixel_HF, means_HF : 64))
-          {
-            means_HF[c] += neighbour_pixel_HF[k][c] / 9.f;
-          }
-
-        // Get the local variance per channel
-        dt_aligned_pixel_t variance_HF = { 0.f, 0.f, 0.f, 0.f };
-        for(size_t k = 0; k < 9; k++)
-          for_each_channel(c, aligned(variance_HF, neighbour_pixel_HF, means_HF))
-          {
-            variance_HF[c] += sqf(neighbour_pixel_HF[k][c] - means_HF[c]) / 9.f;
-          }
-
-        // Find the channel most likely to contain details = max( variance(HF) )
-        size_t guiding_channel_HF = ALPHA;
-        float guiding_value_HF = 0.f;
-        for(size_t c = 0; c < 3; ++c)
-        {
-          if(variance_HF[c] > guiding_value_HF)
-          {
-            guiding_value_HF = variance_HF[c];
-            guiding_channel_HF = c;
-          }
-        }
-
-        // Compute the linear regression channel = f(guide)
-        dt_aligned_pixel_t covariance_HF = { 0.f, 0.f, 0.f, 0.f };
-        for(size_t k = 0; k < 9; k++)
-          for_each_channel(c, aligned(variance_HF, covariance_HF, neighbour_pixel_HF, means_HF))
-          {
-            covariance_HF[c] += (neighbour_pixel_HF[k][c] - means_HF[c])
-                                * (neighbour_pixel_HF[k][guiding_channel_HF] - means_HF[guiding_channel_HF]) / 9.f;
-          }
-
-        const float scale_multiplier = 1.f / radius_sq;
-        const dt_aligned_pixel_t alpha_ch = { clipping_mask[index + RED], clipping_mask[index + GREEN], clipping_mask[index + BLUE], clipping_mask[index + ALPHA] };
-
-        dt_aligned_pixel_t a_HF, b_HF;
-        for_each_channel(c, aligned(out, neighbour_pixel_HF, a_HF, b_HF, covariance_HF, variance_HF, means_HF, high_frequency, alpha_ch))
-        {
-          // Get a and b s.t. y = a * x + b, y = test data, x = guide
-          a_HF[c] = fmaxf(covariance_HF[c] / (variance_HF[guiding_channel_HF]), 0.f);
-          b_HF[c] = means_HF[c] - a_HF[c] * means_HF[guiding_channel_HF];
-
-          high_frequency[c] = alpha_ch[c] * scale_multiplier * (a_HF[c] * high_frequency[guiding_channel_HF] + b_HF[c])
-                            + (1.f - alpha_ch[c] * scale_multiplier) * high_frequency[c];
-        }
-      }
-
-      if(scale & FIRST_SCALE)
-      {
-        // out is not inited yet
-        for_each_channel(c, aligned(out, high_frequency : 64))
-          out[index + c] = high_frequency[c];
-      }
-      else
-      {
-        // just accumulate HF
-        for_each_channel(c, aligned(out, high_frequency : 64))
-          out[index + c] += high_frequency[c];
-      }
-
-      if(scale & LAST_SCALE)
-      {
-        // add the residual and clamp
-        for_each_channel(c, aligned(out, LF, high_frequency : 64))
-          out[index + c] = fmaxf(out[index + c] + LF[index + c], 0.f);
-      }
-
-      // Last step of RGB reconstruct : add noise
-      if((scale & LAST_SCALE) && salt && alpha > 0.f)
-      {
-        // Init random number generator
-        uint32_t DT_ALIGNED_ARRAY state[4] = { splitmix32(j + 1), splitmix32((j + 1) * (i + 3)), splitmix32(1337), splitmix32(666) };
-        xoshiro128plus(state);
-        xoshiro128plus(state);
-        xoshiro128plus(state);
-        xoshiro128plus(state);
-
-        dt_aligned_pixel_t noise = { 0.f };
-        dt_aligned_pixel_t sigma = { 0.20f };
-        const int DT_ALIGNED_ARRAY flip[4] = { TRUE, FALSE, TRUE, FALSE };
-
-        for_each_channel(c,aligned(out, sigma)) sigma[c] = out[index + c] * noise_level;
-
-        // create statistical noise
-        dt_noise_generator_simd(DT_NOISE_POISSONIAN, out + index, sigma, flip, state, noise);
-
-        // Save the noisy interpolated image
-        for_each_channel(c,aligned(out, noise: 64))
-        {
-          // Ensure the noise only brightens the image, since it's clipped
-          noise[c] = out[index + c] + fabsf(noise[c] - out[index + c]);
-          out[index + c] = fmaxf(alpha * noise[c] + alpha_comp * out[index + c], 0.f);
-        }
-      }
-
-      if(scale & LAST_SCALE)
-      {
-        // Break the RGB channels into ratios/norm for the next step of reconstruction
-        const float norm = fmaxf(sqrtf(sqf(out[index + RED]) + sqf(out[index + GREEN]) + sqf(out[index + BLUE])), 1e-6f);
-        for_each_channel(c, aligned(out : 64)) out[index + c] /= norm;
-        out[index + ALPHA] = norm;
-      }
-    }
-  }
-}
-
-static inline void heat_PDE_diffusion(const float *const restrict high_freq,
-                                      const float *const restrict low_freq,
-                                      const float *const restrict clipping_mask,
-                                      float *const restrict output,
-                                      const size_t width,
-                                      const size_t height,
-                                      const int mult,
-                                      const uint8_t scale,
-                                      const float first_order_factor)
-{
-  // Simultaneous inpainting for image structure and texture using anisotropic heat transfer model
-  // https://www.researchgate.net/publication/220663968
-  // modified as follow :
-  //  * apply it in a multi-scale wavelet setup : we basically solve it twice, on the wavelets LF and HF layers.
-  //  * replace the manual texture direction/distance selection by an automatic detection similar to the structure one,
-  //  * generalize the framework for isotropic diffusion and anisotropic weighted on the isophote direction
-  //  * add a variance regularization to better avoid edges.
-  // The sharpness setting mimics the contrast equalizer effect by simply multiplying the HF by some gain.
-
-  float *const restrict out = DT_IS_ALIGNED(output);
-  const float *const restrict LF = DT_IS_ALIGNED(low_freq);
-  const float *const restrict HF = DT_IS_ALIGNED(high_freq);
-
-#ifdef _OPENMP
-#pragma omp parallel for default(none)                                                                            \
-    dt_omp_firstprivate(out, clipping_mask, HF, LF, height, width, mult, scale, first_order_factor) \
-    schedule(static)
-#endif
-  for(size_t row = 0; row < height; ++row)
-  {
-    // interleave the order in which we process the rows so that we minimize cache misses
-    const size_t i = dwt_interleave_rows(row, height, mult);
-    // compute the 'above' and 'below' coordinates, clamping them to the image, once for the entire row
-    const size_t i_neighbours[3]
-      = { MAX((int)(i - mult), (int)0) * width,            // x - mult
-          i * width,                                       // x
-          MIN((int)(i + mult), (int)height - 1) * width }; // x + mult
-    for(size_t j = 0; j < width; ++j)
-    {
-      const size_t idx = (i * width + j);
-      const size_t index = idx * 4;
-
-      // fetch the clipping mask opacity : opaque (alpha = 100 %) where clipped
-      const dt_aligned_pixel_t alpha = { clipping_mask[index + RED],
-                                         clipping_mask[index + GREEN],
-                                         clipping_mask[index + BLUE],
-                                         clipping_mask[index + ALPHA] };
-
-      dt_aligned_pixel_t high_frequency = { HF[index + 0], HF[index + 1], HF[index + 2], HF[index + 3] };
-
-      // The for_each_channel macro uses 4 floats SIMD instructions or 3 float regular ops,
-      // depending on system. Since we don't want to diffuse the norm, make sure to store and restore it later.
-      // This is not much of an issue when processing image at full-res, but more harmful since
-      // we reconstruct highlights on a downscaled variant
-      const float norm_backup = high_frequency[3];
-
-      if(alpha[ALPHA] > 0.f)  // reconstruct
-      {
-        // non-local neighbours coordinates
-        const size_t j_neighbours[3]
-          = { MAX((int)(j - mult), (int)0),           // y - mult
-              j,                                      // y
-              MIN((int)(j + mult), (int)width - 1) }; // y + mult
-
-        // fetch non-local pixels and store them locally and contiguously
-        dt_aligned_pixel_t neighbour_pixel_HF[9];
-        for_four_channels(c, aligned(neighbour_pixel_HF, HF: 16))
-        {
-          neighbour_pixel_HF[3 * 0 + 0][c] = HF[4 * (i_neighbours[0] + j_neighbours[0]) + c];
-          neighbour_pixel_HF[3 * 0 + 1][c] = HF[4 * (i_neighbours[0] + j_neighbours[1]) + c];
-          neighbour_pixel_HF[3 * 0 + 2][c] = HF[4 * (i_neighbours[0] + j_neighbours[2]) + c];
-
-          neighbour_pixel_HF[3 * 1 + 0][c] = HF[4 * (i_neighbours[1] + j_neighbours[0]) + c];
-          neighbour_pixel_HF[3 * 1 + 1][c] = HF[4 * (i_neighbours[1] + j_neighbours[1]) + c];
-          neighbour_pixel_HF[3 * 1 + 2][c] = HF[4 * (i_neighbours[1] + j_neighbours[2]) + c];
-
-          neighbour_pixel_HF[3 * 2 + 0][c] = HF[4 * (i_neighbours[2] + j_neighbours[0]) + c];
-          neighbour_pixel_HF[3 * 2 + 1][c] = HF[4 * (i_neighbours[2] + j_neighbours[1]) + c];
-          neighbour_pixel_HF[3 * 2 + 2][c] = HF[4 * (i_neighbours[2] + j_neighbours[2]) + c];
-        }
-
-        // Compute the laplacian in the direction parallel to the steepest gradient on the norm
-        float DT_ALIGNED_ARRAY anisotropic_kernel_isophote[9] = { 0.25f, 0.5f, 0.25f, 0.5f, -3.f, 0.5f, 0.25f, 0.5f, 0.25f };
-
-        // Convolve the filter to get the laplacian
-        dt_aligned_pixel_t laplacian_HF = { 0.f, 0.f, 0.f, 0.f };
-        for(int k = 0; k < 9; k++)
-        {
-          for_each_channel(c, aligned(laplacian_HF, neighbour_pixel_HF:16) aligned(anisotropic_kernel_isophote: 64))
-            laplacian_HF[c] += neighbour_pixel_HF[k][c] * anisotropic_kernel_isophote[k];
-        }
-
-        // Diffuse
-        const dt_aligned_pixel_t multipliers_HF = { 1.f / B_SPLINE_TO_LAPLACIAN, 1.f / B_SPLINE_TO_LAPLACIAN, 1.f / B_SPLINE_TO_LAPLACIAN, 0.f };
-        for_each_channel(c, aligned(high_frequency, multipliers_HF, laplacian_HF, alpha))
-          high_frequency[c] += alpha[c] * multipliers_HF[c] * (laplacian_HF[c] - first_order_factor * high_frequency[c]);
-
-        // Restore. See above.
-        high_frequency[3] = norm_backup;
-      }
-
-      if((scale & FIRST_SCALE))
-      {
-        // out is not inited yet
-        for_each_channel(c, aligned(out, high_frequency : 64))
-          out[index + c] = high_frequency[c];
-      }
-      else
-      {
-        // just accumulate HF
-        for_each_channel(c, aligned(out, high_frequency : 64))
-          out[index + c] += high_frequency[c];
-      }
-
-      if((scale & LAST_SCALE))
-      {
-        // add the residual and clamp
-        for_each_channel(c, aligned(out, LF, high_frequency : 64))
-          out[index + c] = fmaxf(out[index + c] + LF[index + c], 0.f);
-
-        // renormalize ratios
-        if(alpha[ALPHA] > 0.f)
-        {
-          const float norm = sqrtf(sqf(out[index + RED]) + sqf(out[index + GREEN]) + sqf(out[index + BLUE]));
-          for_each_channel(c, aligned(out, LF, high_frequency : 64))
-            out[index + c] /= (c != ALPHA && norm > 1e-4f) ? norm : 1.f;
-        }
-
-        // Last scale : reconstruct RGB from ratios and norm - norm stays in the 4th channel
-        // we need it to evaluate the gradient
-        for_four_channels(c, aligned(out))
-          out[index + c] = (c == ALPHA) ? out[index + ALPHA] : out[index + c] * out[index + ALPHA];
-      }
-    }
-  }
-}
-
-static inline gint wavelets_process(const float *const restrict in,
-                                    float *const restrict reconstructed,
-                                    const float *const restrict clipping_mask,
-                                    const size_t width,
-                                    const size_t height,
-                                    const int scales,
-                                    float *const restrict HF,
-                                    float *const restrict LF_odd,
-                                    float *const restrict LF_even,
-                                    const diffuse_reconstruct_variant_t variant,
-                                    const float noise_level,
-                                    const int salt,
-                                    const float first_order_factor)
-{
-  gint success = TRUE;
-
-  // À trous decimated wavelet decompose
-  // there is a paper from a guy we know that explains it : https://jo.dreggn.org/home/2010_atrous.pdf
-  // the wavelets decomposition here is the same as the equalizer/atrous module,
-
-  // allocate a one-row temporary buffer for the decomposition
-  size_t padded_size;
-  float *const DT_ALIGNED_ARRAY tempbuf = dt_alloc_perthread_float(4 * width, &padded_size); //TODO: alloc in caller
-  for(int s = 0; s < scales; ++s)
-  {
-    //dt_print(DT_DEBUG_ALWAYS, "CPU Wavelet decompose : scale %i\n", s);
-    const int mult = 1 << s;
-
-    const float *restrict buffer_in;
-    float *restrict buffer_out;
-
-    if(s == 0)
-    {
-      buffer_in = in;
-      buffer_out = LF_odd;
-    }
-    else if(s % 2 != 0)
-    {
-      buffer_in = LF_odd;
-      buffer_out = LF_even;
-    }
-    else
-    {
-      buffer_in = LF_even;
-      buffer_out = LF_odd;
-    }
-
-    decompose_2D_Bspline(buffer_in, HF, buffer_out, width, height, mult, tempbuf, padded_size);
-
-    unsigned int current_scale_type = scale_type(s, scales);
-    const float radius = sqf(equivalent_sigma_at_step(B_SPLINE_SIGMA, s * DS_FACTOR));
-
-    if(variant == DIFFUSE_RECONSTRUCT_RGB)
-      guide_laplacians(HF, buffer_out, clipping_mask, reconstructed, width, height, mult, noise_level, salt, current_scale_type, radius);
-    else
-      heat_PDE_diffusion(HF, buffer_out, clipping_mask, reconstructed, width, height, mult, current_scale_type, first_order_factor);
-
-    if(darktable.dump_pfm_module)
-    {
-      char name[64];
-      sprintf(name, "scale-input-%i", s);
-      dt_dump_pfm(name, buffer_in, width, height,  4 * sizeof(float), "highlights");
-
-      sprintf(name, "scale-blur-%i", s);
-      dt_dump_pfm(name, buffer_out, width, height,  4 * sizeof(float), "highlights");
-    }
-  }
-  dt_free_align(tempbuf);
-
-  return success;
-}
-
-
-static void process_laplacian_bayer(struct dt_iop_module_t *self,
-                                    dt_dev_pixelpipe_iop_t *piece,
-                                    const void *const restrict ivoid,
-                                    void *const restrict ovoid,
-                                    const dt_iop_roi_t *const roi_in,
-                                    const dt_iop_roi_t *const roi_out,
-                                    const dt_aligned_pixel_t clips)
-{
-  dt_iop_highlights_data_t *data = (dt_iop_highlights_data_t *)piece->data;
-
-  const uint32_t filters = piece->pipe->dsc.filters;
-  dt_aligned_pixel_t wb = { 1.f, 1.f, 1.f, 1.f };
-  if(piece->pipe->dsc.temperature.coeffs[0] != 0.f)
-  {
-    wb[0] = piece->pipe->dsc.temperature.coeffs[0];
-    wb[1] = piece->pipe->dsc.temperature.coeffs[1];
-    wb[2] = piece->pipe->dsc.temperature.coeffs[2];
-  }
-
-  const size_t height = roi_in->height;
-  const size_t width = roi_in->width;
-  const size_t ds_height = height / DS_FACTOR;
-  const size_t ds_width = width / DS_FACTOR;
-
-  // [R, G, B, norm] for each pixel
-  float *restrict interpolated, *restrict clipping_mask;
-  // temp buffers for blurs. We will need to cycle between them for memory efficiency
-  float *restrict LF_odd, *restrict LF_even, *restrict temp;
-  // wavelets scales buffers
-  float *restrict HF, *restrict ds_interpolated, *restrict ds_clipping_mask;
-
-  if(!dt_iop_alloc_image_buffers(self, roi_in, roi_out,
-                                 4 | DT_IMGSZ_INPUT, &interpolated,
-                                 4 | DT_IMGSZ_INPUT, &clipping_mask,
-                                 0, NULL))
-  {
-    dt_iop_copy_image_roi(ovoid, ivoid, piece->colors, roi_in, roi_out, 0);
-    return;
-  }
-
-  const dt_iop_roi_t roi_ds = { .x = 0, .y = 0, .height = ds_height, .width = ds_width };
-  if(!dt_iop_alloc_image_buffers(self, &roi_ds, &roi_ds,
-                                 4 | DT_IMGSZ_INPUT, &LF_odd,
-                                 4 | DT_IMGSZ_INPUT, &LF_even,
-                                 4 | DT_IMGSZ_INPUT, &temp,
-                                 4 | DT_IMGSZ_INPUT, &HF,
-                                 4 | DT_IMGSZ_INPUT, &ds_interpolated,
-                                 4 | DT_IMGSZ_INPUT, &ds_clipping_mask,
-                                 0, NULL))
-  {
-    dt_free_align(interpolated);
-    dt_free_align(clipping_mask);
-    dt_iop_copy_image_roi(ovoid, ivoid, piece->colors, roi_in, roi_out, 0);
-    return;
-  }
-
-  const float scale = fmaxf(DS_FACTOR * piece->iscale / (roi_in->scale), 1.f);
-  const float final_radius = (float)((int)(1 << data->scales)) / scale;
-  const int scales = CLAMP((int)ceilf(log2f(final_radius)), 1, MAX_NUM_SCALES);
-
-  const float noise_level = data->noise_level / scale;
-
-  const float *const restrict input = (const float *const restrict)ivoid;
-  float *const restrict output = (float *const restrict)ovoid;
-
-  _interpolate_and_mask(input, interpolated, clipping_mask, clips, wb, filters, width, height);
-  dt_box_mean(clipping_mask, height, width, 4, 2, 1);
-
-  // Downsample
-  interpolate_bilinear(clipping_mask, width, height, ds_clipping_mask, ds_width, ds_height, 4);
-  interpolate_bilinear(interpolated, width, height, ds_interpolated, ds_width, ds_height, 4);
-
-  for(int i = 0; i < data->iterations; i++)
-  {
-    const int salt = (i == data->iterations - 1); // add noise on the last iteration only
-    wavelets_process(ds_interpolated, temp, ds_clipping_mask, ds_width, ds_height, scales, HF, LF_odd,
-                     LF_even, DIFFUSE_RECONSTRUCT_RGB, noise_level, salt, data->solid_color);
-    wavelets_process(temp, ds_interpolated, ds_clipping_mask, ds_width, ds_height, scales, HF, LF_odd,
-                     LF_even, DIFFUSE_RECONSTRUCT_CHROMA, noise_level, salt, data->solid_color);
-  }
-
-  // Upsample
-  interpolate_bilinear(ds_interpolated, ds_width, ds_height, interpolated, width, height, 4);
-  _remosaic_and_replace(input, interpolated, clipping_mask, output, wb, filters, width, height);
-
-  if(darktable.dump_pfm_module)
-  {
-    dt_dump_pfm("interpolated", interpolated, width, height,  4 * sizeof(float), "highlights");
-    dt_dump_pfm("clipping_mask", clipping_mask, width, height,  4 * sizeof(float), "highlights");
-  }
-
-  dt_free_align(interpolated);
-  dt_free_align(clipping_mask);
-  dt_free_align(temp);
-  dt_free_align(LF_even);
-  dt_free_align(LF_odd);
-  dt_free_align(HF);
-  dt_free_align(ds_interpolated);
-  dt_free_align(ds_clipping_mask);
-}
-
-#ifdef HAVE_OPENCL
-static inline cl_int wavelets_process_cl(const int devid,
-                                         cl_mem in,
-                                         cl_mem reconstructed,
-                                         cl_mem clipping_mask,
-                                         const size_t sizes[3],
-                                         const int width,
-                                         const int height,
-                                         dt_iop_highlights_global_data_t *const gd,
-                                         const int scales,
-                                         cl_mem HF,
-                                         cl_mem LF_odd,
-                                         cl_mem LF_even,
-                                         const diffuse_reconstruct_variant_t variant,
-                                         const float noise_level,
-                                         const int salt,
-                                         const float solid_color)
-{
-  cl_int err = DT_OPENCL_DEFAULT_ERROR;
-
-  // À trous wavelet decompose
-  // there is a paper from a guy we know that explains it : https://jo.dreggn.org/home/2010_atrous.pdf
-  // the wavelets decomposition here is the same as the equalizer/atrous module,
-  for(int s = 0; s < scales; ++s)
-  {
-    //dt_print(DT_DEBUG_ALWAYS, "GPU Wavelet decompose : scale %i\n", s);
-    const int mult = 1 << s;
-
-    cl_mem buffer_in;
-    cl_mem buffer_out;
-
-    if(s == 0)
-    {
-      buffer_in = in;
-      buffer_out = LF_odd;
-    }
-    else if(s % 2 != 0)
-    {
-      buffer_in = LF_odd;
-      buffer_out = LF_even;
-    }
-    else
-    {
-      buffer_in = LF_even;
-      buffer_out = LF_odd;
-    }
-
-    // Compute wavelets low-frequency scales
-    dt_opencl_set_kernel_args(devid, gd->kernel_filmic_bspline_horizontal, 0,
-      CLARG(buffer_in), CLARG(HF), CLARG(width), CLARG(height), CLARG(mult));
-    err = dt_opencl_enqueue_kernel_2d(devid, gd->kernel_filmic_bspline_horizontal, sizes);
-    if(err != CL_SUCCESS) return err;
-
-    dt_opencl_set_kernel_args(devid, gd->kernel_filmic_bspline_vertical, 0,
-      CLARG(HF), CLARG(buffer_out), CLARG(width), CLARG(height), CLARG(mult));
-    err = dt_opencl_enqueue_kernel_2d(devid, gd->kernel_filmic_bspline_vertical, sizes);
-    if(err != CL_SUCCESS) return err;
-
-    // Compute wavelets high-frequency scales and backup the maximum of texture over the RGB channels
-    // Note : HF = detail - LF
-    dt_opencl_set_kernel_args(devid, gd->kernel_filmic_wavelets_detail, 0,
-      CLARG(buffer_in), CLARG(buffer_out), CLARG(HF), CLARG(width), CLARG(height));
-    err = dt_opencl_enqueue_kernel_2d(devid, gd->kernel_filmic_wavelets_detail, sizes);
-    if(err != CL_SUCCESS) return err;
-
-    unsigned int current_scale_type = scale_type(s, scales);
-    const float radius = sqf(equivalent_sigma_at_step(B_SPLINE_SIGMA, s * DS_FACTOR));
-
-    // Compute wavelets low-frequency scales
-    if(variant == DIFFUSE_RECONSTRUCT_RGB)
-    {
-      dt_opencl_set_kernel_args(devid, gd->kernel_highlights_guide_laplacians, 0,
-        CLARG(HF), CLARG(buffer_out), CLARG(clipping_mask),
-        CLARG(reconstructed), // read-only
-        CLARG(reconstructed), // write-only
-        CLARG(width), CLARG(height), CLARG(mult), CLARG(noise_level), CLARG(salt), CLARG(current_scale_type), CLARG(radius));
-      err = dt_opencl_enqueue_kernel_2d(devid, gd->kernel_highlights_guide_laplacians, sizes);
-      if(err != CL_SUCCESS) return err;
-    }
-    else // DIFFUSE_RECONSTRUCT_CHROMA
-    {
-      dt_opencl_set_kernel_args(devid, gd->kernel_highlights_diffuse_color, 0,
-        CLARG(HF), CLARG(buffer_out), CLARG(clipping_mask),
-        CLARG(reconstructed), // read-only
-        CLARG(reconstructed), // write-only
-        CLARG(width), CLARG(height), CLARG(mult), CLARG(current_scale_type), CLARG(solid_color));
-      err = dt_opencl_enqueue_kernel_2d(devid, gd->kernel_highlights_diffuse_color, sizes);
-      if(err != CL_SUCCESS) return err;
-    }
-  }
-
-  return err;
-}
-
-static cl_int process_laplacian_bayer_cl(struct dt_iop_module_t *self,
-                                         dt_dev_pixelpipe_iop_t *piece,
-                                         cl_mem dev_in,
-                                         cl_mem dev_out,
-                                         const dt_iop_roi_t *const roi_in,
-                                         const dt_iop_roi_t *const roi_out,
-                                         const dt_aligned_pixel_t clips)
-{
-  dt_iop_highlights_data_t *data = (dt_iop_highlights_data_t *)piece->data;
-  dt_iop_highlights_global_data_t *gd = (dt_iop_highlights_global_data_t *)self->global_data;
-
-  cl_int err = DT_OPENCL_DEFAULT_ERROR;
-
-  const int devid = piece->pipe->devid;
-  const int width = roi_in->width;
-  const int height = roi_in->height;
-
-  const int ds_height = height / DS_FACTOR;
-  const int ds_width = width / DS_FACTOR;
-
-  size_t sizes[] = { ROUNDUPDWD(width, devid), ROUNDUPDHT(height, devid), 1 };
-  size_t ds_sizes[] = { ROUNDUPDWD(ds_width, devid), ROUNDUPDHT(ds_height, devid), 1 };
-
-  const uint32_t filters = piece->pipe->dsc.filters;
-
-  dt_aligned_pixel_t wb = { 1.f, 1.f, 1.f, 1.f };
-  if(piece->pipe->dsc.temperature.coeffs[0] != 0.f)
-  {
-    wb[0] = piece->pipe->dsc.temperature.coeffs[0];
-    wb[1] = piece->pipe->dsc.temperature.coeffs[1];
-    wb[2] = piece->pipe->dsc.temperature.coeffs[2];
-  }
-
-  cl_mem interpolated = dt_opencl_alloc_device(devid, sizes[0], sizes[1], sizeof(float) * 4);  // [R, G, B, norm] for each pixel
-  cl_mem clipping_mask = dt_opencl_alloc_device(devid, sizes[0], sizes[1], sizeof(float) * 4); // [R, G, B, norm] for each pixel
-
-  // temp buffer for blurs. We will need to cycle between them for memory efficiency
-  cl_mem LF_odd = dt_opencl_alloc_device(devid, ds_sizes[0], ds_sizes[1], sizeof(float) * 4);
-  cl_mem LF_even = dt_opencl_alloc_device(devid, ds_sizes[0], ds_sizes[1], sizeof(float) * 4);
-  cl_mem temp = dt_opencl_alloc_device(devid, sizes[0], sizes[1], sizeof(float) * 4); // need full size here for blurring
-
-  const float scale = fmaxf(DS_FACTOR * piece->iscale / (roi_in->scale), 1.f);
-  const float final_radius = (float)((int)(1 << data->scales)) / scale;
-  const int scales = CLAMP((int)ceilf(log2f(final_radius)), 1, MAX_NUM_SCALES);
-
-  const float noise_level = data->noise_level / scale;
-
-  // wavelets scales buffers
-  cl_mem HF = dt_opencl_alloc_device(devid, ds_sizes[0], ds_sizes[1], sizeof(float) * 4);
-  cl_mem ds_interpolated = dt_opencl_alloc_device(devid, ds_sizes[0], ds_sizes[1], sizeof(float) * 4);
-  cl_mem ds_clipping_mask = dt_opencl_alloc_device(devid, ds_sizes[0], ds_sizes[1], sizeof(float) * 4);
-
-  cl_mem clips_cl = dt_opencl_copy_host_to_device_constant(devid, 4 * sizeof(float), (float*)clips);
-  cl_mem wb_cl = dt_opencl_copy_host_to_device_constant(devid, 4 * sizeof(float), (float*)wb);
-
-  dt_opencl_set_kernel_args(devid, gd->kernel_highlights_bilinear_and_mask, 0,
-    CLARG(dev_in), CLARG(interpolated), CLARG(temp),
-    CLARG(clips_cl), CLARG(wb_cl), CLARG(filters), CLARG(roi_out->width), CLARG(roi_out->height));
-  err = dt_opencl_enqueue_kernel_2d(devid, gd->kernel_highlights_bilinear_and_mask, sizes);
-  dt_opencl_release_mem_object(clips_cl);
-  if(err != CL_SUCCESS) goto error;
-
-  dt_opencl_set_kernel_args(devid, gd->kernel_highlights_box_blur, 0, CLARG(temp), CLARG(clipping_mask),
-    CLARG(roi_out->width), CLARG(roi_out->height));
-  err = dt_opencl_enqueue_kernel_2d(devid, gd->kernel_highlights_box_blur, sizes);
-  if(err != CL_SUCCESS) goto error;
-
-  // Downsample
-  const int RGBa = TRUE;
-  dt_opencl_set_kernel_args(devid, gd->kernel_interpolate_bilinear, 0,
-    CLARG(clipping_mask), CLARG(width), CLARG(height), CLARG(ds_clipping_mask), CLARG(ds_width), CLARG(ds_height), CLARG(RGBa));
-  err = dt_opencl_enqueue_kernel_2d(devid, gd->kernel_interpolate_bilinear, ds_sizes);
-  if(err != CL_SUCCESS) goto error;
-
-  dt_opencl_set_kernel_args(devid, gd->kernel_interpolate_bilinear, 0,
-    CLARG(interpolated), CLARG(width), CLARG(height), CLARG(ds_interpolated), CLARG(ds_width), CLARG(ds_height), CLARG(RGBa));
-  err = dt_opencl_enqueue_kernel_2d(devid, gd->kernel_interpolate_bilinear, ds_sizes);
-  if(err != CL_SUCCESS) goto error;
-
-  for(int i = 0; i < data->iterations; i++)
-  {
-    const int salt = (i == data->iterations - 1); // add noise on the last iteration only
-    err = wavelets_process_cl(devid, ds_interpolated, temp, ds_clipping_mask, ds_sizes, ds_width, ds_height, gd, scales, HF,
-                              LF_odd, LF_even, DIFFUSE_RECONSTRUCT_RGB, noise_level, salt, data->solid_color);
-    if(err != CL_SUCCESS) goto error;
-
-    err = wavelets_process_cl(devid, temp, ds_interpolated, ds_clipping_mask, ds_sizes, ds_width, ds_height, gd, scales, HF,
-                              LF_odd, LF_even, DIFFUSE_RECONSTRUCT_CHROMA, noise_level, salt, data->solid_color);
-    if(err != CL_SUCCESS) goto error;
-  }
-
-  // Upsample
-  dt_opencl_set_kernel_args(devid, gd->kernel_interpolate_bilinear, 0,
-    CLARG(ds_interpolated), CLARG(ds_width), CLARG(ds_height), CLARG(interpolated), CLARG(width), CLARG(height));
-  err = dt_opencl_enqueue_kernel_2d(devid, gd->kernel_interpolate_bilinear, sizes);
-  if(err != CL_SUCCESS) goto error;
-
-  // Remosaic
-  dt_opencl_set_kernel_args(devid, gd->kernel_highlights_remosaic_and_replace, 0,
-    CLARG(dev_in), CLARG(interpolated), CLARG(clipping_mask), CLARG(dev_out),
-    CLARG(wb_cl), CLARG(filters), CLARG(width), CLARG(height));
-  err = dt_opencl_enqueue_kernel_2d(devid, gd->kernel_highlights_remosaic_and_replace, sizes);
-  if(err != CL_SUCCESS) goto error;
-
-  // cleanup and exit on success
-  if(wb_cl) dt_opencl_release_mem_object(wb_cl);
-  if(interpolated) dt_opencl_release_mem_object(interpolated);
-  if(clipping_mask) dt_opencl_release_mem_object(clipping_mask);
-  if(temp) dt_opencl_release_mem_object(temp);
-  if(LF_even) dt_opencl_release_mem_object(LF_even);
-  if(LF_odd) dt_opencl_release_mem_object(LF_odd);
-  if(HF) dt_opencl_release_mem_object(HF);
-  dt_opencl_release_mem_object(ds_clipping_mask);
-  dt_opencl_release_mem_object(ds_interpolated);
-  return err;
-
-error:
-  if(wb_cl) dt_opencl_release_mem_object(wb_cl);
-  if(interpolated) dt_opencl_release_mem_object(interpolated);
-  if(clipping_mask) dt_opencl_release_mem_object(clipping_mask);
-  if(temp) dt_opencl_release_mem_object(temp);
-  if(LF_even) dt_opencl_release_mem_object(LF_even);
-  if(LF_odd) dt_opencl_release_mem_object(LF_odd);
-  if(HF) dt_opencl_release_mem_object(HF);
-
-  dt_print(DT_DEBUG_OPENCL, "[opencl_highlights] couldn't enqueue kernel! %s\n", cl_errstr(err));
-  return err;
-}
-#endif
-
-static void process_clip(dt_dev_pixelpipe_iop_t *piece, const void *const ivoid, void *const ovoid,
-                         const dt_iop_roi_t *const roi_in, const dt_iop_roi_t *const roi_out,
+static void process_clip(dt_dev_pixelpipe_iop_t *piece,
+                         const void *const ivoid,
+                         void *const ovoid,
+                         const dt_iop_roi_t *const roi_in,
+                         const dt_iop_roi_t *const roi_out,
                          const float clip)
 {
   const float *const in = (const float *const)ivoid;
@@ -1985,9 +640,12 @@ static void process_clip(dt_dev_pixelpipe_iop_t *piece, const void *const ivoid,
   }
 }
 
-static void process_visualize(dt_dev_pixelpipe_iop_t *piece, const void *const ivoid, void *const ovoid,
-                         const dt_iop_roi_t *const roi_in, const dt_iop_roi_t *const roi_out,
-                         dt_iop_highlights_data_t *data)
+static void process_visualize(dt_dev_pixelpipe_iop_t *piece,
+                              const void *const ivoid,
+                              void *const ovoid,
+                              const dt_iop_roi_t *const roi_in,
+                              const dt_iop_roi_t *const roi_out,
+                              dt_iop_highlights_data_t *data)
 {
   const uint8_t(*const xtrans)[6] = (const uint8_t(*const)[6])piece->pipe->dsc.xtrans;
   const uint32_t filters = piece->pipe->dsc.filters;
@@ -2039,59 +697,12 @@ static void process_visualize(dt_dev_pixelpipe_iop_t *piece, const void *const i
   }
 }
 
-#include "iop/hlrecovery_v2.c"
-#include "iop/opposed.c"
-
-/* inpaint opposed and segmentation based algorithms want the whole image for proper calculation
-   of chrominance correction and best candidates so we change both rois.
-*/
-void modify_roi_out(dt_iop_module_t *self, dt_dev_pixelpipe_iop_t *piece, dt_iop_roi_t *roi_out,
-                    const dt_iop_roi_t *const roi_in)
-{
-  *roi_out = *roi_in;
-  // it can never hurt to make sure
-  roi_out->x = MAX(0, roi_in->x);
-  roi_out->y = MAX(0, roi_in->y);
-}
-
-void modify_roi_in(dt_iop_module_t *self, dt_dev_pixelpipe_iop_t *piece, const dt_iop_roi_t *const roi_out,
-                   dt_iop_roi_t *roi_in)
-{
-  *roi_in = *roi_out;
-
-  dt_iop_highlights_data_t *d = (dt_iop_highlights_data_t *)piece->data;
-  const gboolean use_opposing = (d->mode == DT_IOP_HIGHLIGHTS_OPPOSED) || (d->mode == DT_IOP_HIGHLIGHTS_SEGMENTS);
-  const gboolean fullpipe = piece->pipe->type & DT_DEV_PIXELPIPE_FULL;
-  /* When do we need to expand the roi to maximum of the full input data?
-     1. Certainly not if any other than opposed or the segmentation based algo is used.
-   */
-  if(!use_opposing)
-    return;
-
-  dt_iop_highlights_gui_data_t *g = (dt_iop_highlights_gui_data_t *)self->gui_data;
-  const gboolean clipmask = (g != NULL) ? (g->hlr_mask_mode == DT_HIGHLIGHTS_MASK_CLIPPED) : FALSE;
-  /*
-     2. Certainly not if we show the clipped mask in fullpipe as that is also safe with current roi
-     3. Certainly not for linear raws as they miss the automatic downscaler provided by the demosaicer stage, so
-        the expanding to full image data does not work as we do a downscaling very early in
-        the pixelpipe. So - no quality achieved but really bad performance.
-        See #12998 and #12993 for a lengthy discussion
-  */
-  if((fullpipe && clipmask) || (piece->pipe->dsc.filters == 0))
-    return;
-
-  /* We require the correct (full-image-data) expansion with a defined scale for all pixelpipes for proper
-     aligning and scaling in the demosiacer
-  */
-  roi_in->x = 0;
-  roi_in->y = 0;
-  roi_in->width = piece->buf_in.width;
-  roi_in->height = piece->buf_in.height;
-  roi_in->scale = 1.0f;
-}
-
-void process(struct dt_iop_module_t *self, dt_dev_pixelpipe_iop_t *piece, const void *const ivoid,
-             void *const ovoid, const dt_iop_roi_t *const roi_in, const dt_iop_roi_t *const roi_out)
+void process(struct dt_iop_module_t *self,
+             dt_dev_pixelpipe_iop_t *piece,
+             const void *const ivoid,
+             void *const ovoid,
+             const dt_iop_roi_t *const roi_in,
+             const dt_iop_roi_t *const roi_out)
 {
   const uint32_t filters = piece->pipe->dsc.filters;
   dt_iop_highlights_data_t *data = (dt_iop_highlights_data_t *)piece->data;
@@ -2159,8 +770,7 @@ void process(struct dt_iop_module_t *self, dt_dev_pixelpipe_iop_t *piece, const 
         const uint8_t(*const xtrans)[6] = (const uint8_t(*const)[6])piece->pipe->dsc.xtrans;
 #ifdef _OPENMP
 #pragma omp parallel for default(none) \
-        dt_omp_firstprivate(clips, filters, ivoid, ovoid, roi_in, roi_out, \
-                            xtrans) \
+        dt_omp_firstprivate(clips, filters, ivoid, ovoid, roi_in, roi_out, xtrans) \
         schedule(static)
 #endif
         for(int j = 0; j < roi_out->height; j++)
@@ -2170,8 +780,7 @@ void process(struct dt_iop_module_t *self, dt_dev_pixelpipe_iop_t *piece, const 
         }
 #ifdef _OPENMP
 #pragma omp parallel for default(none) \
-        dt_omp_firstprivate(clips, filters, ivoid, ovoid, roi_in, roi_out, \
-                            xtrans) \
+        dt_omp_firstprivate(clips, filters, ivoid, ovoid, roi_in, roi_out, xtrans) \
         schedule(static)
 #endif
         for(int i = 0; i < roi_out->width; i++)
@@ -2262,7 +871,9 @@ void process(struct dt_iop_module_t *self, dt_dev_pixelpipe_iop_t *piece, const 
   }
 }
 
-void commit_params(struct dt_iop_module_t *self, dt_iop_params_t *p1, dt_dev_pixelpipe_t *pipe,
+void commit_params(struct dt_iop_module_t *self,
+                   dt_iop_params_t *p1,
+                   dt_dev_pixelpipe_t *pipe,
                    dt_dev_pixelpipe_iop_t *piece)
 {
   dt_iop_highlights_params_t *p = (dt_iop_highlights_params_t *)p1;
