+++ src/iop/hlreconstruct/inpaint.c
@@ -1,0 +1,287 @@
+/*
+   This file is part of darktable,
+   Copyright (C) 2010-2023 darktable developers.
+
+   darktable is free software: you can redistribute it and/or modify
+   it under the terms of the GNU General Public License as published by
+   the Free Software Foundation, either version 3 of the License, or
+   (at your option) any later version.
+
+   darktable is distributed in the hope that it will be useful,
+   but WITHOUT ANY WARRANTY; without even the implied warranty of
+   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+   GNU General Public License for more details.
+
+   You should have received a copy of the GNU General Public License
+   along with darktable.  If not, see <http://www.gnu.org/licenses/>.
+*/
+
+/* interpolate value for a pixel, ideal via ratio to nearby pixel */
+static inline float interp_pix_xtrans(const int ratio_next,
+                                      const ssize_t offset_next,
+                                      const float clip0,
+                                      const float clip_next,
+                                      const float *const in,
+                                      const float *const ratios)
+{
+  assert(ratio_next != 0);
+  // it's OK to exceed clipping of current pixel's color based on a
+  // neighbor -- that is the purpose of interpolating highlight
+  // colors
+  const float clip_val = fmaxf(clip0, clip_next);
+  if(in[offset_next] >= clip_next - 1e-5f)
+  {
+    // next pixel is also clipped
+    return clip_val;
+  }
+  else
+  {
+    // set this pixel in ratio to the next
+    assert(ratio_next != 0);
+    if(ratio_next > 0)
+      return fminf(in[offset_next] / ratios[ratio_next], clip_val);
+    else
+      return fminf(in[offset_next] * ratios[-ratio_next], clip_val);
+  }
+}
+
+static inline void interpolate_color_xtrans(const void *const ivoid,
+                                            void *const ovoid,
+                                            const dt_iop_roi_t *const roi_in,
+                                            const dt_iop_roi_t *const roi_out,
+                                            int dim,
+                                            int dir,
+                                            int other,
+                                            const float *const clip,
+                                            const uint8_t (*const xtrans)[6],
+                                            const int pass)
+{
+  // In Bayer each row/col has only green/red or green/blue
+  // transitions, hence can reconstruct color by single ratio per
+  // row. In x-trans there can be transitions between arbitrary colors
+  // in a row/col (and 2x2 green blocks which provide no color
+  // transition information). Hence calculate multiple color ratios
+  // for each row/col.
+
+  // Lookup for color ratios, e.g. red -> blue is roff[0][2] and blue
+  // -> red is roff[2][0]. Returned value is an index into ratios. If
+  // negative, then need to invert the ratio. Identity color
+  // transitions aren't used.
+  const int roff[3][3] = {{ 0, -1, -2},
+                          { 1,  0, -3},
+                          { 2,  3,  0}};
+  // record ratios of color transitions 0:unused, 1:RG, 2:RB, and 3:GB
+  dt_aligned_pixel_t ratios = {1.0f, 1.0f, 1.0f, 1.0f};
+
+  // passes are 0:+x, 1:-x, 2:+y, 3:-y
+  // dims are 0:traverse a row, 1:traverse a column
+  // dir is 1:left to right, -1: right to left
+  int i = (dim == 0) ? 0 : other;
+  int j = (dim == 0) ? other : 0;
+  const ssize_t offs = (ssize_t)(dim ? roi_out->width : 1) * ((dir < 0) ? -1 : 1);
+  const ssize_t offl = offs - (dim ? 1 : roi_out->width);
+  const ssize_t offr = offs + (dim ? 1 : roi_out->width);
+  int beg, end;
+  if(dir == 1)
+  {
+    beg = 0;
+    end = (dim == 0) ? roi_out->width : roi_out->height;
+  }
+  else
+  {
+    beg = ((dim == 0) ? roi_out->width : roi_out->height) - 1;
+    end = -1;
+  }
+
+  float *in, *out;
+  if(dim == 1)
+  {
+    out = (float *)ovoid + (size_t)i + (size_t)beg * roi_out->width;
+    in = (float *)ivoid + (size_t)i + (size_t)beg * roi_in->width;
+  }
+  else
+  {
+    out = (float *)ovoid + (size_t)beg + (size_t)j * roi_out->width;
+    in = (float *)ivoid + (size_t)beg + (size_t)j * roi_in->width;
+  }
+
+  for(int k = beg; k != end; k += dir)
+  {
+    if(dim == 1)
+      j = k;
+    else
+      i = k;
+
+    const uint8_t f0 = FCxtrans(j, i, roi_in, xtrans);
+    const uint8_t f1 = FCxtrans(dim ? (j + dir) : j, dim ? i : (i + dir), roi_in, xtrans);
+    const uint8_t fl = FCxtrans(dim ? (j + dir) : (j - 1), dim ? (i - 1) : (i + dir), roi_in, xtrans);
+    const uint8_t fr = FCxtrans(dim ? (j + dir) : (j + 1), dim ? (i + 1) : (i + dir), roi_in, xtrans);
+    const float clip0 = clip[f0];
+    const float clip1 = clip[f1];
+    const float clipl = clip[fl];
+    const float clipr = clip[fr];
+    const float clip_max = fmaxf(fmaxf(clip[0], clip[1]), clip[2]);
+
+    if(i == 0 || i == roi_out->width - 1 || j == 0 || j == roi_out->height - 1)
+    {
+      if(pass == 3) out[0] = fminf(clip_max, in[0]);
+    }
+    else
+    {
+      // ratio to next pixel if this & next are unclamped and not in
+      // 2x2 green block
+      if((f0 != f1) &&
+          (in[0] < clip0 && in[0] > 1e-5f) &&
+          (in[offs] < clip1 && in[offs] > 1e-5f))
+      {
+        const int r = roff[f0][f1];
+        assert(r != 0);
+        if(r > 0)
+          ratios[r] = (3.f * ratios[r] + (in[offs] / in[0])) / 4.f;
+        else
+          ratios[-r] = (3.f * ratios[-r] + (in[0] / in[offs])) / 4.f;
+      }
+
+      if(in[0] >= clip0 - 1e-5f)
+      {
+        // interplate color for clipped pixel
+        float add;
+        if(f0 != f1)
+          // next pixel is different color
+          add =
+            interp_pix_xtrans(roff[f0][f1], offs, clip0, clip1, in, ratios);
+        else
+          // at start of 2x2 green block, look diagonally
+          add = (fl != f0) ?
+            interp_pix_xtrans(roff[f0][fl], offl, clip0, clipl, in, ratios) :
+            interp_pix_xtrans(roff[f0][fr], offr, clip0, clipr, in, ratios);
+
+        if(pass == 0)
+          out[0] = add;
+        else if(pass == 3)
+          out[0] = fminf(clip_max, (out[0] + add) / 4.0f);
+        else
+          out[0] += add;
+      }
+      else
+      {
+        // pixel is not clipped
+        if(pass == 3) out[0] = in[0];
+      }
+    }
+    out += offs;
+    in += offs;
+  }
+}
+
+static inline void interpolate_color(const void *const ivoid,
+                                     void *const ovoid,
+                                     const dt_iop_roi_t *const roi_out,
+                                     int dim,
+                                     int dir,
+                                     int other,
+                                     const float *clip,
+                                     const uint32_t filters,
+                                     const int pass)
+{
+  float ratio = 1.0f;
+  float *in, *out;
+
+  int i = 0, j = 0;
+  if(dim == 0)
+    j = other;
+  else
+    i = other;
+  ssize_t offs = dim ? roi_out->width : 1;
+  if(dir < 0) offs = -offs;
+  int beg, end;
+  if(dim == 0 && dir == 1)
+  {
+    beg = 0;
+    end = roi_out->width;
+  }
+  else if(dim == 0 && dir == -1)
+  {
+    beg = roi_out->width - 1;
+    end = -1;
+  }
+  else if(dim == 1 && dir == 1)
+  {
+    beg = 0;
+    end = roi_out->height;
+  }
+  else if(dim == 1 && dir == -1)
+  {
+    beg = roi_out->height - 1;
+    end = -1;
+  }
+  else
+    return;
+
+  if(dim == 1)
+  {
+    out = (float *)ovoid + i + (size_t)beg * roi_out->width;
+    in = (float *)ivoid + i + (size_t)beg * roi_out->width;
+  }
+  else
+  {
+    out = (float *)ovoid + beg + (size_t)j * roi_out->width;
+    in = (float *)ivoid + beg + (size_t)j * roi_out->width;
+  }
+  for(int k = beg; k != end; k += dir)
+  {
+    if(dim == 1)
+      j = k;
+    else
+      i = k;
+    const float clip0 = clip[FC(j, i, filters)];
+    const float clip1 = clip[FC(dim ? (j + 1) : j, dim ? i : (i + 1), filters)];
+    if(i == 0 || i == roi_out->width - 1 || j == 0 || j == roi_out->height - 1)
+    {
+      if(pass == 3) out[0] = in[0];
+    }
+    else
+    {
+      if(in[0] < clip0 && in[0] > 1e-5f)
+      { // both are not clipped
+        if(in[offs] < clip1 && in[offs] > 1e-5f)
+        { // update ratio, exponential decay. ratio = in[odd]/in[even]
+          if(k & 1)
+            ratio = (3.0f * ratio + in[0] / in[offs]) / 4.0f;
+          else
+            ratio = (3.0f * ratio + in[offs] / in[0]) / 4.0f;
+        }
+      }
+
+      if(in[0] >= clip0 - 1e-5f)
+      { // in[0] is clipped, restore it as in[1] adjusted according to ratio
+        float add = 0.0f;
+        if(in[offs] >= clip1 - 1e-5f)
+          add = fmaxf(clip0, clip1);
+        else if(k & 1)
+          add = in[offs] * ratio;
+        else
+          add = in[offs] / ratio;
+
+        if(pass == 0)
+          out[0] = add;
+        else if(pass == 3)
+          out[0] = (out[0] + add) / 4.0f;
+        else
+          out[0] += add;
+      }
+      else
+      {
+        if(pass == 3) out[0] = in[0];
+      }
+    }
+    out += offs;
+    in += offs;
+  }
+}
+
+// clang-format off
+// modelines: These editor modelines have been set for all relevant files by tools/update_modelines.py
+// vim: shiftwidth=2 expandtab tabstop=2 cindent
+// kate: tab-indents: off; indent-width 2; replace-tabs on; indent-mode cstyle; remove-trailing-spaces modified;
+// clang-format on
